+++ packages/mission-engine/index.js
@@ -101,6 +101,22 @@ const {
 } = require('./ArtifactBus');
 const ArtifactRegistry = require('./ArtifactRegistry');
 const OperatorArtifactInjection = require('./OperatorArtifactInjection');
+const ArtifactResolver = require('./ArtifactResolver');
+const {
+  parseIntent,
+  classifyUnit,
+  PLAN_CATEGORIES,
+} = require('./IntentParser');
+const {
+  MISSION_PLAN_VERSION,
+  RESERVED_RUNTIME_FIELDS,
+  buildMissionPlan,
+  validateMissionPlan,
+  summarizeMissionPlan,
+  executableObjectiveText,
+  containsOperatorInstructionLeak,
+  resolveExecutionRequest,
+} = require('./MissionPlan');
 
 module.exports = {
   MISSION_STATUS,
@@ -171,6 +187,23 @@ module.exports = {
   createArtifactBus,
   ArtifactRegistry,
   OperatorArtifactInjection,
+  ArtifactResolver,
+  resolveArtifacts: ArtifactResolver.resolveArtifacts,
+  deriveRequiredArtifacts: ArtifactResolver.deriveRequiredArtifacts,
+  acquisitionOptions: ArtifactResolver.acquisitionOptions,
+  ARTIFACT_SOURCES: ArtifactResolver.ARTIFACT_SOURCES,
   detectOperatorProspectListInMessage:
     OperatorArtifactInjection.detectOperatorProspectListInMessage,
+  // SPEC-050 / ADR-034
+  parseIntent,
+  classifyUnit,
+  PLAN_CATEGORIES,
+  MISSION_PLAN_VERSION,
+  RESERVED_RUNTIME_FIELDS,
+  buildMissionPlan,
+  validateMissionPlan,
+  summarizeMissionPlan,
+  executableObjectiveText,
+  containsOperatorInstructionLeak,
+  resolveExecutionRequest,
 };